Refactor telegram service to be separate (#3044)
* Refactor telegram service to be separate

* Fix log
